requirements.txt
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
bcrypt==4.1.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
dnspython==2.6.1
docs==0.1.0
ecdsa==0.19.0
fastapi==0.110.3
h11==0.14.0
idna==3.7
Jinja2==3.1.3
MarkupSafe==2.1.5
outcome==1.3.0.post0
packaging==24.0
passlib==1.7.4
pyasn1==0.6.0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
PyJWT==2.8.0
pymongo==4.7.1
PySocks==1.7.1
python-decouple==3.8
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
requests==2.31.0
rsa==4.9
selenium==4.20.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.37.2
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.29.0
webdriver-manager==4.0.1
wsproto==1.2.0
